+++ CHANGELOG.md
@@ -1,9 +1,13 @@
-## Unrelease
+## [Unreleased]
 
 ### Fixed
+* Fixed background blur and replacement pixelated video on desktop. (Issue https://github.com/aws-samples/amazon-chime-react-native-demo/issues/145)
+* Updated unit tests according to above fix
+* Load model only when frame dimensions change
 * Fixed crash from openGL due to potential race condition.
 
 ## [0.17.1] - 2022-06-03
+
 
 ## [0.17.0] - 2022-05-18